+++ packages/genesys/packages/react-components-core/src/theme/hook/index.ts
@@ -1,2 +1,3 @@
 export { default as useSetTheme } from "./useSetTheme";
 export { default as useTheme } from "./useTheme";
+export { default as useColor } from "./useColor";