+++ packages/genesys/packages/react-native-components/src/display/Label/Label.types.ts
@@ -1,4 +1,4 @@
-import { CoreLabelProps } from "@peersyst/react-components-core";
+import { CoreLabelProps, ThemeColor } from "@peersyst/react-components-core";
 import { TextProps, TextStyle, ViewStyle } from "react-native";
 
 export type LabelStyle = ViewStyle & {
@@ -6,6 +6,10 @@ export type LabelStyle = ViewStyle & {
 };
 
 export interface LabelProps extends CoreLabelProps {
+    /**
+     * Label color
+     */
+    color?: ThemeColor;
     /**
      * Label style
      */