+++ packages/genesys/packages/react-native-components/src/display/Typography/Typography.tsx
@@ -3,20 +3,24 @@ import { useTheme } from "@peersyst/react-native-styled";
 import { TypographyRoot } from "./Typography.styles";
 import { useMergeDefaultProps } from "@peersyst/react-components-core";
 import { useGlobalStyles } from "../../config";
+import { useColorStyle } from "../../theme";
 
 const Typography = (props: TypographyProps): JSX.Element => {
     const {
         variant,
         children,
         style: { light: lightStyleProp = {}, ...styleProp } = {},
+        color: colorProp,
         ...rest
     } = useMergeDefaultProps("Typography", props);
     const { typography } = useTheme();
     const variantStyle = typography[variant];
     const { light: lightGlobalStyle, ...typographyGlobalStyle } = useGlobalStyles("Typography");
 
-    const style = { ...typographyGlobalStyle, ...styleProp };
-    const lightStyle = { ...lightGlobalStyle, ...lightStyleProp };
+    const colorStyle = useColorStyle(colorProp);
+
+    const style = { ...typographyGlobalStyle, ...colorStyle, ...styleProp };
+    const lightStyle = { ...lightGlobalStyle, ...colorStyle, ...lightStyleProp };
 
     return (
         <TypographyRoot