+++ packages/genesys/packages/react-native-components/src/theme/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useColorStyle } from "./useColorStyle";