+++ packages/genesys/packages/react-native-components/src/theme/hooks/useColorStyle.ts
@@ -1,0 +1,6 @@
+import { ThemeColor, useColor } from "@peersyst/react-components-core";
+
+export default function (color: ThemeColor | undefined): { color: string } | {} {
+    const colorValue = useColor(color);
+    return colorValue ? { color: colorValue } : {};
+}